requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-WTF==1.2.1
python-dotenv==1.0.0
Werkzeug==2.3.7
supabase==2.3.4
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0.post1
gunicorn==21.2.0
